Update scala version, add scalacOptions
This fixes a leaking warning trigger by macro expansion.
To prevent poluting the source code with `()`, `-Ywarn-value-discard` is disabled in `boopickle/shared`.

+++ boopickle/js/src/main/scala/boopickle/StringCodec.scala
@@ -79,7 +79,7 @@ object StringCodec extends StringCodecFuncs {
     val ta = new Uint16Array(s.length)
     var i = 0
     while (i < s.length) {
-      ta(i) = s.charAt(i).toShort
+      ta(i) = s.charAt(i).toInt
       i += 1
     }
     TypedArrayBuffer.wrap(new Int8Array(ta.buffer))
